backend/pyproject.toml
[tool.poetry]
name = "mateup league"
version = "0.1.0"
description = ""
authors = ["Anderson <andersonlimacrv@gmail.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"

pydantic = {extras = ["email"], version = "^2.11.7"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
alembic = "^1.16.4"

python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
bcrypt = "^4.3.0"

python-dotenv = "^1.1.0"
loguru = "^0.7.3"
tenacity = "^9.1.2"

uvicorn = "^0.34.3"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"
